Move back to TSDX and make package typesafe
