+++ .lintstagedrc
@@ -1,5 +1,5 @@
 {
-  "*.{js,jsx,json,md,html,css}": [
-    "prettier --write"
+  "*": [
+    "yarn format:all"
   ]
 }
