+++ @types/custom.d.ts
@@ -1,0 +1,7 @@
+declare module '*.svg' {
+  import { SVGFactory } from 'react';
+
+  const component: SVGFactory;
+
+  export default component;
+}
