--- custom.d.ts
@@ -1,6 +1,0 @@
-declare module '*.svg' {
-  import { ReactSVGElement } from 'react';
-
-  const component: ReactSVGElement;
-  export default component;
-}
