+++ package.json
@@ -1,66 +1,63 @@
 {
-  "name": "svg-icons",
+  "name": "react-svg-icons",
   "version": "1.0.0",
   "description": "Componetized SVG Icons for React",
-  "main": "dist/index.production.cjs.js",
-  "module": "dist/index.production.esm.js",
-  "types": "dist/index.d.ts",
-  "author": "Kyle Holmberg",
   "license": "MIT",
-  "keywords": [
-    "react",
-    "svg",
-    "icons",
-    "svg"
+  "main": "dist/index.js",
+  "module": "dist/icons.esm.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
   ],
   "scripts": {
-    "build": "yarn build:dev && yarn build:prod",
-    "build:dev": "yarn rollup -c --environment BUILD:development",
-    "build:prod": "yarn rollup -c --environment BUILD:production",
+    "build": "tsdx build",
     "build:storybook": "build-storybook",
-    "format": "yarn prettier --write \"**/*.{js,json,md,html,css}\" && yarn format:svgs",
+    "format:all": "yarn format:source && yarn format:svgs",
+    "format:source": "yarn prettier --write \"**/*\"",
     "format:svgs": "yarn svgo -f src/svgs --pretty --config=svgo.config.json",
-    "prepublishOnly": "yarn build:prod && yarn test",
-    "storybook": "yarn build:prod && start-storybook",
-    "test": "yarn build:prod && yarn jest"
+    "lint": "tsdx lint",
+    "prepare": "tsdx build",
+    "storybook": "start-storybook -p 6006",
+    "test": "yarn jest --maxWorkers=2",
+    "test:ci": "yarn test --ci=true --coverage=true"
   },
   "peerDependencies": {
     "react": ">=16",
     "react-dom": ">=16"
   },
   "devDependencies": {
+    "@acorns/browserslist-config": "^1.0.0",
     "@babel/core": "^7.8.3",
     "@babel/plugin-transform-react-display-name": "^7.8.3",
     "@babel/preset-env": "^7.8.3",
     "@babel/preset-react": "^7.8.3",
-    "@innocuous/functions": "^2.1.0",
-    "@storybook/addon-knobs": "^5.3.8",
+    "@storybook/addon-knobs": "^5.3.9",
+    "@storybook/addons": "^5.3.8",
     "@storybook/react": "^5.3.8",
-    "babel-jest": "^24.9.0",
+    "@types/jest": "^24.9.1",
+    "@types/react": "^16.9.19",
+    "@types/react-dom": "^16.9.5",
+    "@types/react-test-renderer": "^16.9.2",
+    "babel-jest": "^25.1.0",
     "babel-loader": "^8.0.6",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "enzyme-to-json": "^3.4.3",
-    "husky": "^4.2.0",
-    "jest": "^24.9.0",
+    "husky": "^4.2.1",
+    "jest": "^25.1.0",
     "jest-svg-transformer": "^1.0.0",
-    "lint-staged": "^10.0.1",
-    "path": "^0.12.7",
+    "lint-staged": "^10.0.2",
     "prettier": "^1.19.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "react-svg-loader": "3.0.3",
+    "react-svg-loader": "^3.0.3",
     "react-test-renderer": "^16.12.0",
-    "rollup": "^1.29.1",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-peer-deps-external": "^2.2.0",
-    "rollup-plugin-react-svg": "3.0.3",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-sourcemaps": "^0.5.0",
-    "rollup-plugin-terser": "^5.2.0",
+    "rollup-plugin-react-svg": "^3.0.3",
     "svgo": "^1.3.2",
+    "ts-loader": "^6.2.1",
+    "tsdx": "^0.12.3",
+    "tslib": "^1.10.0",
     "typescript": "^3.7.5"
   }
 }
