+++ src/index.test.js
@@ -3,10 +3,11 @@ import renderer from 'react-test-renderer';
 
 /*
  * We don't want to import the src modules because Jest doesn't have a way to handle SVGs.
- * In this system, we're just rendering React components which happen to render `svgs`. This will
- * more closely emulate the manner with which our consumers will use our components, also.
+ * In this system, we're just rendering React components which happen to render `svgs`.
+ * Additionally, this will more closely emulate the manner with which our consumers will use our
+ * components.
  */
-import * as ProductionBundleIconsMap from '../dist/index.production.esm';
+import * as ProductionBundleIconsMap from '../dist/react-svg-icons.esm';
 
 const createSnapshotTest = (Component) => {
   const tree = renderer.create(Component).toJSON();
