+++ src/index.ts
@@ -1,4 +1,18 @@
-export { default as Alert } from './svgs/alert.svg';
-export { default as Flag } from './svgs/flag.svg';
-export { default as Home } from './svgs/home.svg';
-export { default as MagnifyingGlass } from './svgs/maginifying-glass.svg';
+import { SVGFactory } from 'react';
+
+import _Alert from './svgs/alert.svg';
+import _Flag from './svgs/flag.svg';
+import _Home from './svgs/home.svg';
+import _MagnifyingGlass from './svgs/magnifying-glass.svg';
+
+const Alert: SVGFactory = _Alert;
+const Flag: SVGFactory = _Flag;
+const Home: SVGFactory = _Home;
+const MagnifyingGlass: SVGFactory = _MagnifyingGlass;
+
+export {
+  Alert,
+  Flag,
+  Home,
+  MagnifyingGlass,
+};
